#14 Renamed gfx vertex and added gfx mesh
