+++ code/inc/gfx/element/gfx_vertex.hpp
