+++ code/src/gfx/gfx_vertex.cpp
@@ -29,7 +29,7 @@
 ** 2.  INCLUDE FILES
 **===============================================================================================*/
 
-#include "gfx/element/gfx_element_vertex.hpp"
+#include "gfx/element/gfx_vertex.hpp"
 
 /*=================================================================================================
 ** 3.  DECLARATIONS
